requirements.txt
selenium
beautifulsoup4
requests
Pillow
webdriver-manager

